INT - moved resolver builder to utils for reusability
